Remove libtrust dep from api
