+++ daemon/testdata/keyfile
